Refactor: Move tracing bridge to its own package
So we can call it without creating circular dependencies.
