+++ pkg/util/tracing/tracing_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package mimir
+package tracing
 
 import (
 	"context"
